binary-search-tree: move stub to the right place
